Add cpp analyser

+++ addNumbers.cpp
@@ -1,5 +1,5 @@
 #include "addNumbers.h"
 
-float add(int a, int b) {
+int add(int a, int b) {
     return a + b;
 }
